Change scene class

+++ VulkanScene/Timer.h
@@ -4,13 +4,12 @@
 
 using namespace std::chrono;
 
-// timer class for scene rendering
 class Timer
 {
 public:
 	// returns difference between last time point and current time in seconds
 	// and save current time as last time point
-	double getDeltaSec();
+	float getDeltaSec();
 
 private:
 	time_point<steady_clock> lastTimePoint = (time_point<steady_clock>::max)();
